Move `ConfigurationCacheFingerprint.Header` serialization to `ConfigurationCacheIO`
